refactor(stock): Clean stock items
