+++ src/stock-items/stock-items-table.scss
