+++ src/stock-items/stock-items.component.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import StockItemsTableComponent from "./table/stock-items-table.component";
+import StockItemsTableComponent from "./stock-items-table.component";
 
 const StockItems = () => {
   return <StockItemsTableComponent />;
